@@ -1,44 +1,32 @@
 name: excuses
 description: A new Flutter project.
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  remote_state:
-  rxdart:
-  provider:
-  freezed_annotation:
-  dio: ^3.0.9
-  animations:
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  animations: ^2.0.0
+  cupertino_icons: ^1.0.2
+  dio: ^4.0.0
+  equatable: ^2.0.2
+  remote_state: ^2.0.0
+  routemaster: ^0.8.1
+  rxdart: ^0.27.0
+  provider: ^5.0.0
+  freezed_annotation:
 
 dev_dependencies:
   build_runner: ^1.7.2
   freezed:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^1.0.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
   # The following line ensures that the Material Icons font is
