@@ -2,9 +2,9 @@ import 'package:excuses/models/models.dart';
 import 'package:flutter/material.dart';
 
 class ExcuseCard extends StatelessWidget {
-  final Excuse excuse;
+  final Excuse? excuse;
 
-  const ExcuseCard._({Key key, this.excuse}) : super(key: key);
+  const ExcuseCard._({Key? key, this.excuse}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -12,8 +12,8 @@ class ExcuseCard extends StatelessWidget {
 
     return Center(
       child: Text(
-        excuse.text,
-        style: textTheme.display3,
+        excuse!.text!,
+        style: textTheme.headline2,
       ),
     );
   }
