@@ -4,19 +4,19 @@ import 'package:excuses/ui/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 
 class ExcusePageView extends StatelessWidget {
-  final List<Excuse> excuses;
-  final int currentPage;
+  final List<Excuse>? excuses;
+  final int? currentExcuse;
 
-  const ExcusePageView({Key key, this.excuses, this.currentPage})
+  const ExcusePageView({Key? key, this.excuses, this.currentExcuse})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    var excuseCards =
-        excuses.map((excuse) => ExcuseCard.fromExcuse(excuse)).toList();
+    var selectedExcuse =
+        excuses!.firstWhere((element) => element.id == currentExcuse);
 
     return PageTransitionSwitcher(
-      duration: Duration(milliseconds: 500),
+      duration: const Duration(milliseconds: 500),
       transitionBuilder: (child, animation, secondAnimation) {
         return SharedAxisTransition(
           child: child,
@@ -25,7 +25,7 @@ class ExcusePageView extends StatelessWidget {
           transitionType: SharedAxisTransitionType.horizontal,
         );
       },
-      child: excuseCards[currentPage],
+      child: ExcuseCard.fromExcuse(selectedExcuse),
     );
   }
 }
